backend/requirements.txt
Flask
psycopg2-binary
python-dotenv
groq
streamlit
flask-cors
requests

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6